reorganize maptoolcircle test files
