--- tests/src/app/maptoolcircle/testqgsmaptoolcircle.cpp
@@ -1,230 +1,0 @@
-/***************************************************************************
-     testqgsmaptoolcircle.cpp
-     ------------------------
-    Date                 : April 2021
-    Copyright            : (C) 2021 by Loïc Bartoletti
-    Email                : loic dot bartoletti @oslandia dot com
- ***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
-
-#include "qgstest.h"
-
-#include "qgisapp.h"
-#include "qgsgeometry.h"
-#include "qgsmapcanvas.h"
-#include "qgssettingsregistrycore.h"
-#include "qgsvectorlayer.h"
-#include "qgsmaptooladdfeature.h"
-
-#include "testqgsmaptoolutils.h"
-#include "qgsmaptoolcircle2points.h"
-#include "qgsmaptoolcircle3points.h"
-#include "qgsmaptoolcirclecenterpoint.h"
-
-
-class TestQgsMapToolCircle : public QObject
-{
-    Q_OBJECT
-
-  public:
-    TestQgsMapToolCircle();
-
-  private slots:
-    void initTestCase();
-    void cleanupTestCase();
-
-    void testCircleFrom2Points();
-    void testCircleFrom2PointsWithDeletedVertex();
-    void testCircleFrom3Points();
-    void testCircleFrom3PointsWithDeletedVertex();
-    void testCircleFromCenterPoint();
-    void testCircleFromCenterPointWithDeletedVertex();
-
-  private:
-    QgisApp *mQgisApp = nullptr;
-    QgsMapToolCapture *mParentTool = nullptr;
-    QgsMapCanvas *mCanvas = nullptr;
-    QgsVectorLayer *mLayer = nullptr;
-};
-
-TestQgsMapToolCircle::TestQgsMapToolCircle() = default;
-
-
-//runs before all tests
-void TestQgsMapToolCircle::initTestCase()
-{
-  QgsApplication::init();
-  QgsApplication::initQgis();
-
-  mQgisApp = new QgisApp();
-
-  mCanvas = new QgsMapCanvas();
-  mCanvas->setDestinationCrs( QgsCoordinateReferenceSystem( QStringLiteral( "EPSG:27700" ) ) );
-
-  // make testing layers
-  mLayer = new QgsVectorLayer( QStringLiteral( "LineString?crs=EPSG:27700" ), QStringLiteral( "layer line " ), QStringLiteral( "memory" ) );
-  QVERIFY( mLayer->isValid() );
-  QgsProject::instance()->addMapLayers( QList<QgsMapLayer *>() << mLayer );
-
-  // set layers in canvas
-  mCanvas->setLayers( QList<QgsMapLayer *>() << mLayer );
-  mCanvas->setCurrentLayer( mLayer );
-
-  mParentTool = new QgsMapToolAddFeature( mCanvas, QgsMapToolCapture::CaptureLine );
-}
-
-void TestQgsMapToolCircle::cleanupTestCase()
-{
-  QgsApplication::exitQgis();
-}
-
-void TestQgsMapToolCircle::testCircleFrom2Points()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircle2Points mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseMove( 0, 2 );
-  utils.mouseClick( 0, 2, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 1 1, 0 0, -1 1, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-void TestQgsMapToolCircle::testCircleFrom2PointsWithDeletedVertex()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircle2Points mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-
-  utils.mouseClick( 4, 1, Qt::LeftButton );
-  utils.keyClick( Qt::Key_Backspace );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseMove( 0, 2 );
-  utils.mouseClick( 0, 2, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 1 1, 0 0, -1 1, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-void TestQgsMapToolCircle::testCircleFrom3Points()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircle3Points mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseClick( 0, 2, Qt::LeftButton );
-  utils.mouseMove( 1, 1 );
-  utils.mouseClick( 1, 1, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 1 1, 0 0, -1 1, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-void TestQgsMapToolCircle::testCircleFrom3PointsWithDeletedVertex()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircle3Points mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseClick( 4, 1, Qt::LeftButton );
-  utils.keyClick( Qt::Key_Backspace );
-  utils.mouseClick( 0, 2, Qt::LeftButton );
-  utils.mouseMove( 1, 1 );
-  utils.mouseClick( 1, 1, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 1 1, 0 0, -1 1, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-void TestQgsMapToolCircle::testCircleFromCenterPoint()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircleCenterPoint mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseMove( 0, 2 );
-  utils.mouseClick( 0, 2, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 2 0, 0 -2, -2 0, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-void TestQgsMapToolCircle::testCircleFromCenterPointWithDeletedVertex()
-{
-  mLayer->startEditing();
-
-  QgsMapToolCircleCenterPoint mapTool( mParentTool, mCanvas );
-  mCanvas->setMapTool( &mapTool );
-
-  TestQgsMapToolAdvancedDigitizingUtils utils( &mapTool );
-
-  utils.mouseClick( 4, 1, Qt::LeftButton );
-  utils.keyClick( Qt::Key_Backspace );
-  utils.mouseClick( 0, 0, Qt::LeftButton );
-  utils.mouseMove( 0, 2 );
-  utils.mouseClick( 0, 2, Qt::RightButton );
-  QgsFeatureId newFid = utils.newFeatureId();
-
-  QCOMPARE( mLayer->featureCount(), ( long )1 );
-  QgsFeature f = mLayer->getFeature( newFid );
-
-  QString wkt = "CompoundCurve (CircularString (0 2, 2 0, 0 -2, -2 0, 0 2))";
-  QCOMPARE( f.geometry().asWkt(), wkt );
-
-  mLayer->rollBack();
-}
-
-QGSTEST_MAIN( TestQgsMapToolCircle )
-#include "testqgsmaptoolcircle.moc"
